% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_line.R
\name{add_service_line}
\alias{add_service_line}
\title{Add provider service and service line to dataset}
\usage{
add_service_line(data)
}
\arguments{
\item{data}{data.frame and tibble containing hospital records}
}
\value{
The same dataset as input, but with one new variable
  (\code{ServiceLine}) and one updated variable (\code{ProviderService}).
}
\description{
Based on a function developed by SHIPS after discussions with health records,
the code implemented below adds provider service and service line information
to the dataset.
}
\details{
This function requires four fields to be present in the data:
\code{PatientServiceSubService}, \code{ProviderServiceDesc},
\code{ProviderService}, and \code{ProviderNumber}.
}
