% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_holidays_df.R
\name{create_df_holidays}
\alias{create_df_holidays}
\title{Create data frame containing holiday dates}
\usage{
create_df_holidays(begin, end)
}
\arguments{
\item{begin}{Date indicating when we should start looking for holidays.}

\item{end}{Date indicating when we should stop looking for holidays.}
}
\value{
A data frame with two columns: ds and holiday. The former contains a
  date, and the latter, a description of the holiday.
}
\description{
This is is used with the prophet package to do forecasting of time series.
Adding holiday information may improve the accuracy of the forecast.
}
\examples{
create_df_holidays(lubridate::ymd("20180101"), lubridate::ymd("20181231"))
}
